requirements.txt
requests
beautifulsoup4
python-whois
python-dotenv
reportlab